requirements.txt
python-telegram-bot
SunoAI